// src/api/generated/models/propertyRoomRoomType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * devit-api
 * Devit API
 * OpenAPI spec version: 0.0.1
 */

export type PropertyRoomRoomType = typeof PropertyRoomRoomType[keyof typeof PropertyRoomRoomType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PropertyRoomRoomType = {
  kitchen: 'kitchen',
  balcony: 'balcony',
  terrace: 'terrace',
  garden: 'garden',
  land: 'land',
  garage: 'garage',
  covered_parking: 'covered_parking',
  uncovered_parking: 'uncovered_parking',
  motorcycle_spot: 'motorcycle_spot',
  camper_spot: 'camper_spot',
  boat_spot: 'boat_spot',
  cellar: 'cellar',
  warehouse: 'warehouse',
  storage: 'storage',
  attic: 'attic',
  loft: 'loft',
  carriage_house: 'carriage_house',
  stable: 'stable',
  veranda: 'veranda',
} as const;